+++ NEWS.md
@@ -1,5 +1,6 @@
 # officer.pptx 0.1.1
 
+* remove `add_slide()` patch as now integrated into `{officer}`.
 * remove `phs_with()`. Ported to `{officer}` as of version `0.6.8`.
 * `slide_preview` to plot a slide, `slide_to_png` to save a slide to file. Experimental version `slide_preview_2` 
   leveraging the `doconv` package, but with some problems (#5).